+++ src/dgisim/character/character.py
@@ -55,6 +55,7 @@ __all__ = [
     "Keqing",
     "Klee",
     "KujouSara",
+    "Layla",
     "Lisa",
     "Lyney",
     "MaguuKenki",
@@ -1452,11 +1453,6 @@ class FatuiCryoCicinMage(Character):
 
     @override
     def _skill2(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
-        if sm.CryoCicinsSummon in game_state.get_player(source.pid).summons:
-            cryo_cicins = game_state.get_player(source.pid).summons.just_find(sm.CryoCicinsSummon)
-            cicin_update = replace(cryo_cicins, usages=2)
-        else:
-            cicin_update = sm.CryoCicinsSummon()
         return (
             eft.ReferredDamageEffect(
                 source=source,
@@ -1465,9 +1461,9 @@ class FatuiCryoCicinMage(Character):
                 damage=1,
                 damage_type=DamageType(elemental_skill=True),
             ),
-            eft.UpdateSummonEffect(
+            eft.AddSummonEffect(
                 target_pid=source.pid,
-                summon=cicin_update,
+                summon=sm.CryoCicinsSummon,
             ),
         )
 
@@ -2562,6 +2558,77 @@ class KujouSara(Character):
             eft.AddSummonEffect(
                 target_pid=source.pid,
                 summon=sm.TenguJuuraiStormclusterSummon,
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=2,
+            hiddens=stts.Statuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class Layla(Character):
+    _ELEMENT = Element.CRYO
+    _WEAPON_TYPE = WeaponType.SWORD
+    _TALENT_STATUS = stt.LightsRemitStatus
+    _FACTIONS = frozenset((Faction.SUMERU,))
+
+    _SKILL1_COST = AbstractDice({
+        Element.CRYO: 1,
+        Element.ANY: 2,
+    })
+    _SKILL2_COST = AbstractDice({
+        Element.CRYO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDice({
+        Element.CRYO: 3,
+    })
+
+    def _skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    def _skill2(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.CurtainOfSlumberStatus,
+            ),
+            eft.AddCombatStatusEffect(
+                target_pid=source.pid,
+                status=stt.ShootingStarStatus,
+            ),
+        )
+
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.max_energy,
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.CRYO,
+                damage=3,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+            eft.AddSummonEffect(
+                target_pid=source.pid,
+                summon=sm.CelestialDreamsphereSummon,
             ),
         )